backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
distlib==0.3.7
Django==4.1.13
djongo==1.3.6
dnspython
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
MarkupSafe==2.1.4
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pattern==3.6
pdfminer.six==20221105
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==3.11.0
portend==3.2.0
preshed==3.0.9
protobuf==4.23.4
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==3.12.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
setuptools==69.0.3
smart-open==6.4.0
sqlparse==0.2.4
srsly==2.4.8
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.6
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.4
zc.lockfile==3.0.post1
lime
shap
datasets
pyjwt
django-cors-headers
djangorestframework
torch
spacy>=3.7.2,<3.8.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
sentence-transformers>=1.1.0
pattern>=3.6.0
pymysql
munch>=2.5.0
zss
scipy
nltk